'modal_delete'

+++ blog/views/comment_views.py
@@ -6,8 +6,6 @@ from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 from django.views.generic import CreateView
 from django.contrib.auth.mixins import LoginRequiredMixin
-
-
 @login_required
 def comment_create(request,post_pk):
     post = get_object_or_404(Post,pk = post_pk)
@@ -76,10 +74,15 @@ def comment_edit(request,comment_pk):
 @login_required
 def comment_delete(request,comment_pk):
     comment = get_object_or_404(Comment,pk=comment_pk)
+
+    if comment.author != request.user:
+        messages.error(request,'작성자만삭제가능')
+        return redirect(comment.post)
+
     if request.method == 'POST':
         comment.delete()
         messages.success(request,'포스팅삭제')
         return redirect(comment.post)
-    return render(request,'blog/comment_modal.html',{'comment':comment,})
+    return render(request,'blog/comment_delete.html',{'comment':comment,})
